Create ConfigMap before running the app
