+++ java/java.lsp.server/vscode/src/extension.ts
@@ -400,12 +400,16 @@ export function activate(context: ExtensionContext): VSNetBeansAPI {
     );
     context.subscriptions.push(vscode.commands.registerCommand('cloud.assets.config.create', async function (viewItem) {
             const CONFIG_LOCAL = 'Open a preview of the config in the editor';
-            const CONFIG_TO_CM = 'Upload the config to a ConfigMap artifact in an OCI DevOps Project';
-            const selected: any = await window.showQuickPick([CONFIG_LOCAL, CONFIG_TO_CM], { placeHolder: 'Select a target for the config' });
-            if (selected == CONFIG_TO_CM) {
-                await commands.executeCommand('nbls.cloud.assets.configmap.upload');
+            const CONFIG_TO_DEVOPS_CM = 'Upload the config to a ConfigMap artifact whithin an OCI DevOps Project';
+            const CONFIG_TO_OKE_CM = 'Upload the config to a ConfigMap artifact whithin OKE cluster';
+            const selected: any = await window.showQuickPick([CONFIG_LOCAL, CONFIG_TO_OKE_CM, CONFIG_TO_DEVOPS_CM], { placeHolder: 'Select a target for the config' });
+            if (selected == CONFIG_TO_DEVOPS_CM) {
+                await commands.executeCommand('nbls.cloud.assets.configmap.devops.upload');
                 return;
-            } 
+            } else if (selected == CONFIG_TO_OKE_CM) {
+                await commands.executeCommand('nbls.cloud.assets.configmap.upload');
+                return;               
+            }
             const content = await vscode.commands.executeCommand('nbls.cloud.assets.config.create.local') as string;
             const document = vscode.Uri.parse(`${scheme}:application.properties?${encodeURIComponent(content)}`);
             vscode.workspace.openTextDocument(document).then(doc => {
